requirements.txt
Django==4.0.3
psycopg2==2.9.3
django-allauth==0.50.0
Pillow==9.1.0
pytest-django==4.5.2
djangorestframework==3.13.1
django-rest-swagger==2.2.0
django-rest-auth==0.9.5
djangorestframework-simplejwt==5.1.0
invoke==1.7.0
django-cron==0.6.0
django-environ==0.8.1
django-weekday-field==1.1.0
